requirements.txt
annotated-types==0.7.0
anyio==4.4.0
backoff==2.2.1
boto3==1.34.145
botocore==1.34.145
certifi==2024.7.4
charset-normalizer==3.3.2
cohere==5.6.1
cohere-haystack==1.1.3
distro==1.9.0
dnspython==2.6.1
fastavro==1.9.5
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
haystack-ai==2.3.0
haystack-experimental==0.1.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
lazy-imports==0.3.1
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
ollama-haystack==0.0.7
openai==1.35.15
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
posthog==3.5.0
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
tenacity==8.5.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
types-requests==2.32.0.20240712
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
